projects/book-library-api-tests/requirements.txt
pytest
requests
faker
pydantic
allure-pytest
pytest-xdist
pytest-rerunfailures
jsonschema
python-dotenv
pytest-mock
json-server
